refactor : UserName vo로 변경
