+++ src/main/java/gmbs/controller/Controller.java
@@ -5,6 +5,7 @@ import gmbs.dto.ResultDTO;
 import gmbs.model.*;
 import gmbs.model.players.Player;
 import gmbs.model.vo.UserInput;
+import gmbs.model.vo.UserName;
 import gmbs.view.Input;
 import gmbs.view.Output;
