+++ src/main/java/gmbs/model/BlackJackManager.java
@@ -4,6 +4,7 @@ import gmbs.model.first_class_collection.CardHand;
 import gmbs.model.players.Dealer;
 import gmbs.model.players.Player;
 import gmbs.model.players.User;
+import gmbs.model.vo.UserName;
 
 public class BlackJackManager {
