+++ src/test/java/gmbs/model/WinnerCheckerTest.java
@@ -8,6 +8,7 @@ import gmbs.model.players.Dealer;
 import gmbs.model.players.User;
 import gmbs.model.vo.BlackJackResult;
 import gmbs.model.vo.Card;
+import gmbs.model.vo.UserName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
